% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{assign_matrix_dense}
\alias{assign_matrix_dense}
\title{Create a matrix from a triplet representation}
\usage{
assign_matrix_dense(raw, row.ids, col.ids)
}
\arguments{
\item{raw}{data.frame in triplet form where each row represents a triplet}

\item{row.ids}{The row ids}

\item{col.ids}{The col ids}
}
\value{
A matrix
}
\description{
A somewhat optimized constructor of dense matrices from triplets
}

