% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{ct}
\alias{ct}
\title{Perform the conjugate transpose of a matrix}
\arguments{
\item{m}{A matrix}
}
\value{
THe conjugate transpose of the original matrix
}
\description{
Convenience function
}
\section{Usage}{

ct(m) \%::\% matrix : matrix

ct(m)
}

\section{Details}{

This is a convenience function to compute the conjugate transpose. For
real-valued matrices, ct(m) = t(m).
}
\examples{
x <- matrix(rcomp(16), nrow=4)
ct(x)
}

