% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoff.R
\name{cutoff}
\alias{MaximumLikelihoodFit}
\alias{RandomMatrixFilter}
\alias{cutoff}
\title{Calculate the upper bound of the noise spectrum}
\arguments{
\item{p}{A random matrix}
}
\value{
The eigenvalue associated with the upper bound of the noise
spectrum
}
\description{
Finds the cutoff point between the noise portion of the eigenvalue
spectrum and the "signal" part.
}
\section{Usage}{

cutoff(p) %::% matrix : numeric

cutoff(p, es, estimator) %::% matrix : list : RandomMatrixFilter : numeric
}
\examples{
cutoff(rmatrix(WignerModel(100)))
}

