% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{dmatrix}
\alias{dmatrix}
\title{Get the density for the given random matrix model}
\arguments{
\item{x}{A numeric vector}

\item{model}{The random matrix model}
}
\value{
A vector of corresponding densities
}
\description{
This is the analog to dnorm except for a random matrix
}
\section{Usage}{

dmatrix(x, model) %::% a : WignerModel : a

dmatrix(x, model) %::% a : WishartModel : a

dmatrix(x, model) %::% a : JacobiModel : a
}
\examples{
m <- WignerModel(100)
dmatrix(seq(-1,1,by=0.02), m)
}

