% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{domain}
\alias{domain}
\title{Get the bounds of the eigenvalues for the given model}
\arguments{
\item{model}{The random matrix model}
}
\value{
A vector containing the lower and upper bounds
}
\description{
Currently only supports Wishart matrices
}
\section{Usage}{

domain(model) %::% WishartModel : numeric
}

