% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoff.R
\name{fit.density}
\alias{fit.density}
\title{Fit the eigenvalue spectrum to model}
\arguments{
\item{lambda}{Eigenvalues}

\item{fitter}{A fit function}
}
\value{
Optimal parameters
}
\description{
This only works for fitting the Marcenko-Pastur distribution. It's
also not designed for external use.
}
\examples{
\dontrun{
m <- rmatrix(WishartModel(50, 200))
es <- eigen(m)
fit.density(es, MaximumLikelihoodFit(hint=c(1,1)))
}
}

