% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{print.Ensemble}
\alias{print.Ensemble}
\title{Print a random matrix ensemble}
\usage{
\method{print}{Ensemble}(x, ...)
}
\arguments{
\item{x}{An ensemble of random matrices}

\item{...}{Reserved for later}
}
\value{
Used for side-effects
}
\description{
Pretty print the ensemble instead of dumping a bunch of matrices
}
\seealso{
\code{\link{Ensemble}}
}

