% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{rcomp}
\alias{rcomp}
\title{Generate random complex numbers}
\arguments{
\item{n}{Length of the output vector}

\item{dist}{The distribution for the random number genertor}
}
\value{
A vector of random numbers
}
\description{
Generate random complex numbers using the specified distribution.
By default \code{rnorm} is used.
}
\section{Usage}{

rcomp(n, dist) %::% numeric : Function : complex

rcomp(n, dist=rnorm)
}

\section{Details}{

This function is used primarily to generate random matrices.
}
\examples{
rcomp(10)
rcomp(10, runif)
}

